Added user list visual - actions & tests to do then
