+++ frontend/public/root/pages/services/manage/sub-components/UsersList.scss
@@ -1,0 +1,67 @@
+@import "../../../../variables.scss";
+
+.user__item {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  gap: 15px;
+  padding: 10px;
+  border: 1px solid $_2k25-background-hue-50-transparent;
+  border-radius: 8px;
+  transition: background-color 0.2s ease-in-out;
+  cursor: pointer;
+  &:hover {
+    background-color: $_2k25-background-hue-darker-2-transparent;
+  }
+  .user__item__avatar {
+    width: 40px;
+    height: 40px;
+    border-radius: 50%;
+    object-fit: cover;
+  }
+
+  justify-content: space-between;
+
+  .user__item__info {
+    display: flex;
+    flex-direction: row;
+    gap: 15px;
+    .item__info__header {
+      display: flex;
+      flex-direction: column;
+      .user__item__username {
+        font-weight: 600;
+        font-size: 16px;
+      }
+      .user__item__email {
+        font-size: 14px;
+        opacity: 0.8;
+      }
+    }
+  }
+  .user__item__rights {
+    display: flex;
+    flex-direction: row;
+    gap: 8px;
+    margin-top: 4px;
+    max-width: 50%;
+    min-width: 50%;
+    flex-wrap: wrap;
+    .user__item__right {
+      padding: 2px 6px;
+      border-radius: 4px;
+      font-weight: 500;
+    }
+  }
+
+  .user__item__dates {
+    display: flex;
+    flex-direction: column;
+    font-size: 12px;
+    opacity: 0.7;
+    text-align: right;
+    .user__item__date {
+      margin-top: 2px;
+    }
+  }
+}
