+++ frontend/src/pages/services/manage/pages/overview/components/users/index.tsx
@@ -1,21 +1,12 @@
-import { useEffect, useState } from "react";
-import type { ServicesForUserProps } from "../../../../../../../types/ServicesForUserProps";
+import {  useState } from "react";
+import type { ServicesForUserProps, ServiceUser } from "../../../../../../../types/ServicesForUserProps";
 import type { ServiceOverviewTabs } from "../../types/tabs.type";
 import Loader from "../../../../../../../global/components/Loader";
-import axios from "axios";
+import '../../../../../../../../public/root/pages/services/manage/sub-components/UsersList.scss';
+import useFetchUserList from "./scripts/fetch-user-list";
+import ListedUser from "./components/user";
 
-interface ServiceUser {
-    id: string;
-    username: string;
-    email: string;
-    profile_picture: string;
-    rights: {
-        id: string;
-        name: string;
-        hue: string;
-        description: string;
-    }[];
-}
+
 
 const ServiceUsers = ({
     service,
@@ -31,36 +22,7 @@ const ServiceUsers = ({
 
     const [errorMessage, setErrorMessage] = useState<string | null>(null);
 
-    useEffect(() => {
-        const fetchUsers = async () => {
-            if (service) {
-                axios.post(`${process.env.DUMMY_API_URL_DEV}/user/secure/service/users`, {
-                    service_id: service.id
-                }, {
-                    withCredentials: true
-                }).then((response) => {
-                    if (response.data.success) {
-                        setUsers(response.data.data.serviceUsers || []);
-                    }
-                }).catch((error) => {
-                    console.error("Error fetching service users:", error);
-                    const code = error.response?.data?.code;
-                    if (code) {
-                        // Handle specific error codes if needed
-                        if (code === 403) {
-                            setErrorMessage("You do not have permission to view the users of this service.");
-                        } else {
-                            setErrorMessage(error.response?.data?.message || "An error occurred while fetching service users.");
-                        }
-                    }
-                }).finally(() => {
-                    setLoading(false);
-                });
-            }
-        };
-
-        fetchUsers();
-    }, [service]);
+    useFetchUserList(service ? service.id : "", setUsers, setLoading, setErrorMessage);
 
 
     return (
@@ -77,20 +39,7 @@ const ServiceUsers = ({
                 ) : (
                     users.length > 0 ? (
                         users.map((user) => (
-                            <div key={user.id} className="user__item">
-                                <img src={user.profile_picture} alt={user.username} className="user__item__avatar" />
-                                <div className="user__item__info">
-                                    <span className="user__item__username">{user.username}</span>
-                                    <span className="user__item__email">{user.email}</span>
-                                    <div className="user__item__rights">
-                                        {user.rights.map((right) => (
-                                            <span key={right.id} className="user__item__right" style={{ backgroundColor: `hsl(${right.hue}, 70%, 80%)`, color: `hsl(${right.hue}, 70%, 30%)` }}>
-                                                {right.name}
-                                            </span>
-                                        ))}
-                                    </div>
-                                </div>
-                            </div>
+                            <ListedUser key={user.id} user={user} />
                         ))
                     ) : (
                         <span>
